Move Time Until Rebase and other metrics into header.
